requirements.txt
dash
dash-bootstrap-components
pandas
plotly
databricks-sql-connector
databricks-sdk==0.57.0
python-dotenv
dash-ag-grid
mlflow
